Applied ShaKrune's recommendations for appropriately using his data model. Working code state, can send admin commands via text console. "Show Object" command results sucessfully parsed and displayed.
