+++ Meridian59.DebugUI/ObjectEditor.cs
@@ -6,6 +6,7 @@ using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Windows.Forms;
+using Meridian59.Data.Models.AdminData;
 
 namespace Meridian59.DebugUI
 {
