+++ Meridian59/Meridian59.csproj
@@ -111,6 +111,7 @@
     <Compile Include="Common\Util.cs" />
     <Compile Include="Data\Lists\MailList.cs" />
     <Compile Include="Data\Lists\RoomObjectListFiltered.cs" />
+    <Compile Include="Data\Models\AdminData\AdminData.cs" />
     <Compile Include="Data\Models\CharCreationInfo.cs" />
     <Compile Include="Data\Models\ChatCommand\ChatCommandCast.cs" />
     <Compile Include="Data\Models\ChatCommand\ChatCommandDM.cs" />
@@ -322,6 +323,9 @@
     <Compile Include="Native\Linux.cs" />
     <Compile Include="Native\Windows.cs" />
     <Compile Include="Native\Wrapper.cs" />
+    <Compile Include="Data\Models\AdminData\AdminObject.cs" />
+    <Compile Include="Data\Models\AdminData\AdminObjectProperty.cs" />
+    <Compile Include="Properties\Annotations.cs" />
     <Compile Include="Protocol\GameMessages\GameMode\CancelOfferMessage.cs" />
     <Compile Include="Protocol\GameMessages\GameMode\CounterOfferedMessage.cs" />
     <Compile Include="Protocol\GameMessages\GameMode\ChangePasswordMessage.cs" />
@@ -561,7 +565,9 @@
     <Content Include="zlib\license.txt" />
     <None Include="Resources\Glow.png" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Protocol\GameMessages\AdminMode\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
